Code improvement on master novel view

+++ rental/src/rental/MasterNovel.java
@@ -557,25 +557,30 @@ public class MasterNovel extends javax.swing.JFrame {
      */
     private void destroy(){
         try{
-            if(NovelRepository.destroy(Integer.parseInt(txtId.getText()))){
-                LogRepository.create(new Log(0, authUser.getId(), "Member", "DELETE"));
-                
-                JOptionPane.showMessageDialog(null, "Delete successful", "Delete", JOptionPane.INFORMATION_MESSAGE);
+            int dialogButton = JOptionPane.YES_NO_OPTION;
+            int dialogResult = JOptionPane.showConfirmDialog (null, "Are you sure want to delete this record?","Delete",dialogButton);
+            
+            if(dialogResult == JOptionPane.YES_OPTION){
+                if(NovelRepository.destroy(Integer.parseInt(txtId.getText()))){
+                    LogRepository.create(new Log(0, authUser.getId(), "Member", "DELETE"));
 
-                txtId.setText("");
-                txtTitle.setText("");
-                txtAuthor.setText("");
-                txtLanguage.setText("");
-                txtPublisher.setText("");
-                datePubDate.setCalendar(null);
-                txtRentPrice.setText("");
-                txtPages.setText("");
-                txtIsbn.setText("");
+                    JOptionPane.showMessageDialog(null, "Delete successful", "Delete", JOptionPane.INFORMATION_MESSAGE);
 
-                getNovels();
-            }
-            else{
-                JOptionPane.showMessageDialog(null, "Oops, something's wrong", "Delete", JOptionPane.INFORMATION_MESSAGE);
+                    txtId.setText("");
+                    txtTitle.setText("");
+                    txtAuthor.setText("");
+                    txtLanguage.setText("");
+                    txtPublisher.setText("");
+                    datePubDate.setCalendar(null);
+                    txtRentPrice.setText("");
+                    txtPages.setText("");
+                    txtIsbn.setText("");
+
+                    getNovels();
+                }
+                else{
+                    JOptionPane.showMessageDialog(null, "Oops, something's wrong", "Delete", JOptionPane.INFORMATION_MESSAGE);
+                }
             }
         } catch (Exception e){
             JOptionPane.showMessageDialog(null, "Please choose a record", "Delete", JOptionPane.INFORMATION_MESSAGE);
